Split Main.ts into focused graph modules
